requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
requests==2.31.0
python-socketio==5.9.0
python-engineio==4.7.1
openai==1.3.7
Pillow==10.0.1
numpy==1.24.3
httpx==0.27.2
python-dotenv==1.0.0
pydantic==1.10.12